+++ CHANGELOG.md
@@ -2,3 +2,5 @@
 
 * Added a changelog
 * Refactored to support stem 1.0.0
+* Added codeception
+* Refactored to support Rhubarb 1.0.0